moved first 3 numba functions
